configuring credentials
