+++ .env.local.example
@@ -10,3 +10,5 @@ AUTH_GITHUB_SECRET=
 
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
+
+DATABASE_URL=
