+++ package.json
@@ -19,6 +19,8 @@
     "test:e2e:run": "cypress run"
   },
   "dependencies": {
+    "@next-auth/prisma-adapter": "^1.0.7",
+    "@prisma/client": "^5.20.0",
     "buffer": "^6.0.3",
     "csv-writer": "^1.6.0",
     "cypress-social-logins": "^1.14.2",
@@ -42,6 +44,7 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "postcss": "^8",
+    "prisma": "^5.20.0",
     "tailwindcss": "^3.4.1",
     "ts-node": "^10.9.2",
     "typescript": "^5"
