+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,1 +1,1 @@
-export {GET, POST} from "@/app/auth";
+export {GET, POST} from "@/lib/auth";
