+++ src/lib/prisma.ts
@@ -1,0 +1,7 @@
+import {PrismaClient} from "@prisma/client";
+
+const globalForPrisma = globalThis as unknown as {prisma: PrismaClient};
+
+export const prisma = globalForPrisma.prisma || new PrismaClient();
+
+if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
